update bpk-svgs to use esm
